Create Carly's Clippers
I am the Data Analyst at Carly’s Clippers, the newest hair salon on the block. My job is to go through the lists of data that have been collected in the past couple of weeks. I will be calculating some important metrics that Carly can use to plan out the operation of the business for the rest of the month.